Began modularizing the test process.
At this point, tests are loaded from the config, have their own
execution method within the Test object, are scheduled properly,
and have enough functional code to actually make the HTTP(S) call to
the specified URL and return a response object. At this time, no
analysis is performed on the response data, that still needs to be
written.