Resolved issue #2
Broke the single monolithic test function that had previously existed
out into multiple functions to make a bit more sense from both a
code and logic perspective. Test functions are now stored in a
separate file, tests.js. I'd tried to make the invocation of test
functions dynamic so someone can just add a new type value to a
condition and the code will do the right thing (so long as the
correctly-named function exists somewhere) but I think I was trying
to force reflection on a language that doesn't really use it. Or I
just failed to Google properly.

Lastly, I changed the unit of time for the pollingInterval config
parameter from millisecond to second in the config file because we
don't need that type of resolution and I think will be simpler for
people to work with.